refactor: change location of handler implementations
